% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tobi.R
\name{tobi}
\alias{tobi}
\title{Predicts Stock Price Movement for Given Stock Symbol}
\usage{
tobi(x)
}
\arguments{
\item{symbol}{}
}
\description{
This package predicts whether the stock price at tommorow's market close would be higher or lower compared to today's closing place.
}
\examples{
 stock_predict('AAPL')

}
